test out base model object
